+++ crc.h
@@ -8,6 +8,6 @@
 #ifndef _crc_H_
 #define _crc_H_
 
-unsigned short crc_16(unsigned short crc, unsigned char *data, unsigned int len);
+unsigned short crc_16(unsigned short crc, unsigned char *data, short len);
 
 #endif /* _crc_H_ */
